CRUD Manage Product v2
Semua fitur manage product admin done.

+++ controller/manageProductController.js
@@ -5,18 +5,70 @@ const fs = require('fs')
 module.exports = {
 	getAllProducts: async (req, res) => {
 		// filter queries
-		let { search, minPrice, maxPrice } = req.query
+		let { search, minPrice, maxPrice, sortBy, category, offset } = req.query
 		// get initial products with stock
-		let sql = `select p.id as productId, productName, price, invStock, appStock from products p join stock s on p.id = s.productId`
-		if(search || minPrice || maxPrice){
-			sql += ` where`
-			if(search) sql += ` productName like '%${req.query.search}%'`
-			if(search && minPrice) sql += ` and`
-			if(minPrice) sql += ` price >= ${parseInt(minPrice)}`
-			if((search && maxPrice) || (minPrice && maxPrice)) sql += ` and`
-			if(maxPrice) sql += ` price <= ${parseInt(maxPrice)}`
+		// let sql = `select p.id as productId, productName, price, invStock, appStock from products p join stock s on p.id = s.productId where active = 1`
+		let sql = `select 
+		p.id as productId, 
+		productName, 
+		price, 
+		invStock, 
+		appStock, 
+		IfNull(views, 0) as views, 
+		IfNull(totalPurchased, 0) as totalPurchased,
+		max(case when rn = 1 then category end) category1,
+		max(case when rn = 2 then category end) category2,
+		max(case when rn = 3 then category end) category3
+							from products p
+							left join (
+									select productId, sum(qty)  as totalPurchased
+											from cart
+											group by productId
+							) cart on p.id = cart.productId
+							left join stock s on p.id = s.productId
+							left join (
+									select productId, count(id) as views
+											from product_view
+											group by productId
+							) product_view on p.id = product_view.productId
+							left join (
+									select 
+										prod.id,
+										active,
+										c.category,
+										row_number() over(partition by prod.id order by c.id) rn
+									from products prod
+									join product_category pc on pc.productId = prod.id
+									join category c on c.id = pc.categoryId
+							) as pivot on pivot.id = p.id
+							where p.active = 1 `
+
+		if (search || minPrice || maxPrice) {
+			if (search) sql += ` and productName like '%${req.query.search}%'`
+			if (minPrice) sql += ` and price >= ${parseInt(minPrice)}`
+			if (maxPrice) sql += ` and price <= ${parseInt(maxPrice)}`
 		}
+		sql += ` group by p.id`
+		if (category) {
+			category = JSON.parse(category)
+			for (var i = 0; i < 3; i++) {
+				if (category[i]) {
+					if (i !== 0) sql += ` and`
+					if (i === 0) sql += ` having`
+					sql += ` (max(case when rn = 1 then category end) = '${category[i].label}' or max(case when rn = 2 then category end) = '${category[i].label}' or max(case when rn = 3 then category end) = '${category[i].label}')`
+				}
+			}
+		}
+		if (sortBy) sql += ` order by ${decodeURI(sortBy)}`
 		try {
+			// count total active products without offset to get total
+			let active = await query(sql)
+			let totalActiveProducts = active.length
+			// start fetch product
+			if (offset >= 0) {
+				sql += ` limit 5 offset ${offset}`
+			}
+			console.log(sql)
 			let results = await query(sql)
 			// join images
 			for (i = 0; i < results.length; i++) {
@@ -30,15 +82,13 @@ module.exports = {
 				let category = await query(sql)
 				results[i].category = category
 			}
-			// retrieve categories
-			sql = `select id as value, category as label from category`
-			let categoryList = await query(sql)
 			res.status(200).send({
 				status: 'Success',
-				data: {results, categoryList},
-				message: 'Successfully fetched product'
+				data: { results, totalActiveProducts },
+				message: 'Successfully fetched product',
 			})
 		} catch (err) {
+			console.log(err)
 			res.status(500).send({
 				status: 'Failed',
 				message: err.message,
@@ -51,6 +101,9 @@ module.exports = {
 		const upload = uploader(path, 'PROD').fields([
 			{ name: 'image1' },
 			{ name: 'image2' },
+			{ name: 'image3' },
+			{ name: 'image4' },
+			{ name: 'image5' },
 		])
 		upload(req, res, async (err) => {
 			if (err) {
@@ -59,14 +112,7 @@ module.exports = {
 					message: err.message,
 				})
 			}
-			let {
-				productName,
-				price,
-				invStock,
-				appStock,
-				category1,
-				category2,
-			} = req.body
+			let { productName, price, invStock, appStock, category1, category2, category3 } = req.body
 			let sql = `insert into products (productName, price) values ('${productName}', ${price})`
 			try {
 				// add new product
@@ -74,40 +120,45 @@ module.exports = {
 				// add stock
 				sql = `insert into stock (productId, invStock, appStock) values (${insert.insertId}, ${invStock}, ${appStock})`
 				await query(sql)
-				// add 2 images to product_image
-				req.files = [req.files.image1, req.files.image2]
+				// add 5 images to product_image
+				req.files = [req.files.image1, req.files.image2, req.files.image3, req.files.image4, req.files.image5]
 				req.files.forEach(async (img) => {
-					const imagePath = img ? `${path}/${img[0].filename}` : null
-					sql = `insert into product_image (imagePath, productId) values ('${imagePath}', ${insert.insertId})`
-					try {
-						await query(sql)
-					} catch (err) {
-						fs.unlinkSync(`./public${imagePath}`)
-						res.status(500).send({
-							status: 'Failed',
-							message: err.message,
-						})
+					if (img) {
+						const imagePath = img ? `${path}/${img[0].filename}` : null
+						sql = `insert into product_image (imagePath, productId) values ('${imagePath}', ${insert.insertId})`
+						try {
+							await query(sql)
+						} catch (err) {
+							fs.unlinkSync(`./public${imagePath}`)
+							res.status(500).send({
+								status: 'Failed',
+								message: err.message,
+							})
+						}
 					}
 				})
 				// add categories to product_category
-				let category = [category1, category2]
+				let category = [category1, category2, category3]
 				category.forEach(async (val) => {
-					sql = `insert into product_category (categoryId, productId) values (${val}, ${insert.insertId})`
-					try {
-						await query(sql)
-					} catch (err) {
-						res.status(500).send({
-							status: 'Failed',
-							message: err.message,
-						})
+					if (val) {
+						sql = `insert into product_category (categoryId, productId) values (${val}, ${insert.insertId})`
+						try {
+							await query(sql)
+						} catch (err) {
+							res.status(500).send({
+								status: 'Failed',
+								message: err.message,
+							})
+						}
 					}
 				})
 				res.status(200).send({
 					status: 'Success',
 					data: insert,
-					message: 'Successfully added new product'
+					message: 'Successfully added new product',
 				})
 			} catch (err) {
+				console.log(err)
 				res.status(500).send({
 					status: 'Failed',
 					message: err.message,
@@ -120,7 +171,10 @@ module.exports = {
 		let sql = `select * from product_image where productId = '${id}'`
 		try {
 			let results = await query(sql)
-			let oldImagePath = [results[0].imagePath, results[1].imagePath]
+			let oldImagePath = []
+			results.forEach((image) => {
+				oldImagePath.push(image.imagePath)
+			})
 			sql = `select * from product_category where productId = '${id}'`
 			let categories = await query(sql)
 
@@ -129,6 +183,9 @@ module.exports = {
 			const upload = uploader(path, 'PROD').fields([
 				{ name: 'image1' },
 				{ name: 'image2' },
+				{ name: 'image3' },
+				{ name: 'image4' },
+				{ name: 'image5' },
 			])
 			upload(req, res, async (err) => {
 				if (err) {
@@ -137,67 +194,87 @@ module.exports = {
 						message: err.message,
 					})
 				}
-				let {
-					productName,
-					price,
-					invStock,
-					appStock,
-					category1,
-					category2,
-				} = req.body
-				let sql = `update products set productName = '${productName}', price = ${parseInt(
-					price
-				)} where id = ${id};`
+				let { productName, price, invStock, appStock, category1, category2, category3, deleteImageArr } = req.body
+				let sql = `update products set productName = '${productName}', price = ${parseInt(price)} where id = ${id};`
 				try {
 					// update product table
 					await query(sql)
 					// update stock table
-					sql = `update stock set invStock = ${parseInt(
-						invStock
-					)}, appStock = ${parseInt(
-						appStock
-					)} where productId = ${id};`
+					sql = `update stock set invStock = ${parseInt(invStock)}, appStock = ${parseInt(appStock)} where productId = ${id};`
 					await query(sql)
 					// update product image
-					req.files = [req.files.image1, req.files.image2]
+					req.files = [req.files.image1, req.files.image2, req.files.image3, req.files.image4, req.files.image5]
 					req.files.forEach(async (img, i) => {
-						const imagePath = img
-							? `${path}/${img[0].filename}`
-							: oldImagePath[i]
-						sql = `update product_image set imagePath = '${imagePath}' where id = ${results[i].id};`
-						try {
-							if (img) {
-								await query(sql)
-								fs.unlinkSync(`./public${oldImagePath[i]}`)
+						if (img) {
+							const imagePath = img ? `${path}/${img[0].filename}` : oldImagePath[i]
+							try {
+								if (results[i]) {
+									sql = `update product_image set imagePath = '${imagePath}' where id = ${results[i].id};`
+									await query(sql)
+									if (oldImagePath[i]) {
+										fs.unlinkSync(`./public${oldImagePath[i]}`)
+									}
+								} else {
+									sql = `insert into product_image (imagePath, productId) VALUES ('${imagePath}', '${id}');`
+									await query(sql)
+								}
+							} catch (err) {
+								console.log(err)
+								fs.unlinkSync(`./public${imagePath}`)
+								res.status(500).send({
+									status: 'Failed',
+									message: err.message,
+								})
 							}
-						} catch (err) {
-							fs.unlinkSync(`./public${imagePath}`)
-							res.status(500).send({
-								status: 'Failed',
-								message: err.message,
-							})
 						}
 					})
 					// update category table
-					let category = [category1, category2]
+					let category = [parseInt(category1), parseInt(category2), parseInt(category3)]
 					category.forEach(async (val, i) => {
 						try {
 							if (val) {
-								sql = `update product_category set categoryId = ${val} where id = ${categories[i].id};`
+								if (categories[i] && val !== categories[i].categoryId) {
+									sql = `update product_category set categoryId = ${val} where id = ${categories[i].id};`
+									await query(sql)
+								} else if (!categories[i]) {
+									sql = `insert into product_category (categoryId, productId) values (${val}, ${id})`
+									await query(sql)
+								}
+							} else if (isNaN(val) && categories[i]) {
+								sql = `delete from product_category where id = ${categories[i].id};`
 								await query(sql)
 							}
 						} catch (err) {
+							console.log(err)
 							res.status(500).send({
 								status: 'Failed',
 								message: err.message,
 							})
 						}
 					})
+
+					// delete image
+					deleteImageArr.forEach((id, index) => {
+						if (id !== 'false') {
+							console.log('masuk delete', id)
+							db.query(`delete from product_image where id = ${id}`, (err, results) => {
+								if (err) {
+									return res.status(500).send({
+										status: 'Failed',
+										message: err.message,
+									})
+								}
+								fs.unlinkSync(`./public${oldImagePath[index]}`)
+							})
+						}
+					})
+
 					res.status(200).send({
 						status: 'Success',
 						data: `Successfully updated product with the id of ${id}`,
 					})
 				} catch (err) {
+					console.log(err)
 					res.status(500).send({
 						status: 'Failed',
 						message: err.message,
@@ -212,17 +289,86 @@ module.exports = {
 		}
 	},
 	deleteProduct: async (req, res) => {
-		let sql = `select * from product_image where productId = ${req.params.id}`
+		// let sql = `select * from product_image where productId = ${req.params.id}`
+		// try {
+		// 	let results = await query(sql)
+		// 	results.forEach((img) => {
+		// 		fs.unlinkSync(`./public${img.imagePath}`)
+		// 	})
+		// 	sql = `delete from products where id = ${req.params.id}`
+		// 	await query(sql)
+		// 	res.status(200).send({
+		// 		status: 'Success',
+		// 		message: `Successfully delete product with the id of ${req.params.id}`,
+		// 	})
+		// } catch (err) {
+		// 	res.status(500).send({
+		// 		status: 'Failed',
+		// 		message: err.message,
+		// 	})
+		// }
+		let sql = `update products set active = 0 where id = ${req.params.id}`
 		try {
-			let results = await query(sql)
-			results.forEach((img) => {
-				fs.unlinkSync(`./public${img.imagePath}`)
-			})
-			sql = `delete from products where id = ${req.params.id}`
 			await query(sql)
 			res.status(200).send({
 				status: 'Success',
-				data: `Successfully delete product with the id of ${req.params.id}`,
+				message: `Successfully delete product with the id of ${req.params.id}`,
+			})
+		} catch (err) {
+			res.status(500).send({
+				status: 'Failed',
+				message: err.message,
+			})
+		}
+	},
+	getOneProduct: async (req, res) => {
+		let sql = `select p.id as productId, productName, price, invStock, appStock, active from products p join stock s on p.id = s.productId where p.id = ${req.params.id}`
+		try {
+			let result = await query(sql)
+			sql = `select id as productImageId, imagePath from product_image where productId = ${result[0].productId}`
+			let productImage = await query(sql)
+			result[0].images = productImage
+			// join category
+			sql = `select pc.id as productCategoryId, categoryId, category from product_category pc join category c on c.id = pc.categoryId where productId = ${result[0].productId}`
+			let category = await query(sql)
+			result[0].category = category
+			res.status(200).send({
+				status: 'Success',
+				data: result[0],
+				message: `Successfully fetch product with the id of ${req.params.id}`,
+			})
+		} catch (err) {
+			console.log(err)
+			res.status(500).send({
+				status: 'Failed',
+				message: err.message,
+			})
+		}
+	},
+	fetchCategory: async (req, res) => {
+		let sql = `select id as value, category as label from category`
+		try {
+			let categoryList = await query(sql)
+			res.status(200).send({
+				status: 'Success',
+				data: categoryList,
+				message: 'Successfully fetched category',
+			})
+		} catch (err) {
+			res.status(500).send({
+				status: 'Failed',
+				message: err.message,
+			})
+		}
+	},
+	addCategory: async (req, res) => {
+		let sql = `insert into category (category) values ('${req.body.category}')`
+		try {
+			let insert = await query(sql)
+			res.status(200).send({
+				status: 'Success',
+				data: insert,
+				message: 'Successfully added new category',
 			})
 		} catch (err) {
 			res.status(500).send({
